Minor code restructuring
